feat(BundleHandler): Send heartbeat to keep bundles alive

+++ src/DIRAC/WorkloadManagementSystem/Agent/BundleManagerAgent.py
@@ -4,10 +4,11 @@ from datetime import datetime, timedelta, timezone
 from DIRAC import S_ERROR, S_OK, gConfig
 from DIRAC.Core.Base.AgentModule import AgentModule
 from DIRAC.Core.Utilities.ObjectLoader import ObjectLoader
-from DIRAC.WorkloadManagementSystem.Client import JobStatus, PilotStatus
+from DIRAC.WorkloadManagementSystem.Client import PilotStatus, JobStatus
 from DIRAC.WorkloadManagementSystem.Client.BundlerClient import BundlerClient
 from DIRAC.WorkloadManagementSystem.Client.JobMonitoringClient import JobMonitoringClient
 from DIRAC.WorkloadManagementSystem.DB.BundleDB import BundleDB
+from DIRAC.WorkloadManagementSystem.DB.JobDB import JobDB
 
 
 class BundleManagerAgent(AgentModule):
@@ -17,11 +18,13 @@ class BundleManagerAgent(AgentModule):
         super().__init__(agentName, loadName, baseAgentName, properties)
 
         self.bundleDB = None
+        self.jobDB = None
 
     #############################################################################
 
     def initialize(self):
         self.bundleDB = BundleDB()
+        self.jobDB = JobDB()
         self.jobMonitor = JobMonitoringClient()
         self.bundler = BundlerClient()
         self.maxMinsInBundle = self.am_getOption("MaxMinutesInBundle", defaultValue=10)
@@ -43,6 +46,8 @@ class BundleManagerAgent(AgentModule):
         if not result["OK"]:
             self.log.warn(f"Failed to delete the inputs: {result}")
 
+        self._checkHeartBeat()
+
         return S_OK()
 
     def finalize(self):
@@ -56,7 +61,7 @@ class BundleManagerAgent(AgentModule):
             return result
 
         bundleIDs = result["Value"]
-        self.log.debug(f"> Found {len(bundleIDs)} finished and unpurged bundles")
+        self.log.verbose(f"> Found {len(bundleIDs)} finished and unpurged bundles")
 
         for bundleId in bundleIDs:
             success = True
@@ -67,7 +72,7 @@ class BundleManagerAgent(AgentModule):
 
             jobIDs = result["Value"]
 
-            self.log.debug(f"> Purging inputs of bundle with ID '{bundleId}'")
+            self.log.verbose(f"> Purging inputs of bundle with ID '{bundleId}'")
 
             for jobId in jobIDs:
                 result = self.bundleDB.removeJobInputs(jobId)
@@ -95,7 +100,7 @@ class BundleManagerAgent(AgentModule):
         for bundleInfo in bundles:
             bundleId = bundleInfo["BundleID"]
 
-            result = self.bundleDB.getJobsOfBundle(bundleId)
+            result = self.bundleDB.getJobsOfBundle(bundleId, noInputs=True)
             if not result["OK"]:
                 self.log.error(f"Failed to get the jobs of the bundle '{bundleId}'")
                 return result
@@ -125,6 +130,10 @@ class BundleManagerAgent(AgentModule):
                 if status == JobStatus.KILLED:
                     self.log.info(f"> Status of job '{diracId}' is 'Killed', adding it to the deletion list")
                     killedJobs.append(diracIdToJobId[diracId])
+
+        if not killedJobs:
+            self.log.verbose("Nothing to delete...")
+            return S_OK()
 
         result = self.bundleDB.removeJobsFromBundle(killedJobs)
         if not result["OK"]:
@@ -166,3 +175,23 @@ class BundleManagerAgent(AgentModule):
             result = self.bundler.forceSubmitBundles(bundleIds)
 
         return S_OK()
+
+    def _checkHeartBeat(self):
+        """Hack to avoid stalled jobs when they are not"""
+        self.log.info("Sending heartbeats to running bundles")
+        result = self.bundleDB.getRunningBundles()
+        if not result["OK"]:
+            return result
+
+        for bundleInfo in result["Value"]:
+            if bundleInfo["Status"] == PilotStatus.RUNNING:
+                result = self.bundleDB.getJobsOfBundle(bundleInfo["BundleID"], noInputs=True)
+                if not result["OK"]:
+                    continue
+
+                for _, jobDesc in result["Value"].items():
+                    diracId = jobDesc["DiracID"]
+                    if not diracId:
+                        continue
+
+                    self.jobDB.setHeartBeatData(diracId, {})
